# web/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "EmailChallengeRequest": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "maxLength": 254,
            "minLength": 3,
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "EmailChallengeRequest",
        "type": "object"
      },
      "EmailSessionResponse": {
        "additionalProperties": false,
        "properties": {
          "authenticated": {
            "default": true,
            "title": "Authenticated",
            "type": "boolean"
          },
          "expires_at": {
            "title": "Expires At",
            "type": "string"
          },
          "login_channel": {
            "const": "email",
            "default": "email",
            "title": "Login Channel",
            "type": "string"
          },
          "tenant": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Tenant",
            "type": "object"
          }
        },
        "required": [
          "expires_at",
          "tenant"
        ],
        "title": "EmailSessionResponse",
        "type": "object"
      },
      "EmailVerifyRequest": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "maxLength": 6,
            "minLength": 6,
            "title": "Code",
            "type": "string"
          },
          "email": {
            "maxLength": 254,
            "minLength": 3,
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email",
          "code"
        ],
        "title": "EmailVerifyRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "in": "cookie",
        "name": "__Host-kb_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Notebook Agent Web API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/challenges": {
      "post": {
        "operationId": "request_challenge_api_v1_auth_challenges_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailChallengeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request Challenge",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/session": {
      "delete": {
        "operationId": "logout_api_v1_auth_session_delete",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Logout",
        "tags": [
          "authentication"
        ]
      },
      "get": {
        "operationId": "current_session_api_v1_auth_session_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Current Session",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/verify": {
      "post": {
        "operationId": "verify_api_v1_auth_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify",
        "tags": [
          "authentication"
        ]
      }
    },
...
  }
}